Kernel: Move SB16 to Audio subdirectory
