+++ Kernel/Devices/Audio/SB16.cpp
@@ -9,8 +9,8 @@
 #include <Kernel/Arch/x86/IO.h>
 #include <Kernel/Arch/x86/InterruptDisabler.h>
 #include <Kernel/Debug.h>
+#include <Kernel/Devices/Audio/SB16.h>
 #include <Kernel/Devices/DeviceManagement.h>
-#include <Kernel/Devices/SB16.h>
 #include <Kernel/Memory/AnonymousVMObject.h>
 #include <Kernel/Memory/MemoryManager.h>
 #include <Kernel/Sections.h>
